requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
python-multipart==0.0.6
langchain==0.0.240
openai==0.27.8
PyPDF2==3.0.1
streamlit==1.26.0
requests==2.31.0
python-dotenv==1.0.0